Add interaction system, including a selection mechanic as a working PoC
Interactions work in a state-like manner which is defined through different contexts, with `SelectionContext` being the default for the most of time. This and future contexts are to be inherited from `InteractionContext`. For detailed info, please inspect the mentioned classes.

Side note: Existing player ship movement functionality is broken per this commit and has to be re-implemented as a `ShipMovementContext` or the like.